Feature auction crud

+++ src/main/resources/schema.sql
@@ -3,8 +3,8 @@ DROP TABLE IF EXISTS auction;
 CREATE TABLE IF NOT EXISTS auction (
     id INT NOT NULL UNIQUE AUTO_INCREMENT,
     name VARCHAR(255) NOT NULL,
-    initial_price DOUBLE NOT NULL,
-    current_price DOUBLE,
+    initial_price DECIMAL(6, 2) NOT NULL,
+    current_price DECIMAL(6, 2),
     description VARCHAR(5000),
     end_time DATETIME NOT NULL
 );
